requirements.txt
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1.2
flask>=3.0.0
click>=8.1.7
colorama>=0.4.6
PyJWT>=2.8.0
